Install fixed version of rust
We were previously installing the latest available Rust version
with each container build. In order to have reproducible builds,
the Rust version needs to be fixed and manually updated. This will
ensure that we don't have builds on master fail for the various
crates that already use this container that are due to a change
in the compiler.

Fixes: https://github.com/rust-vmm/rust-vmm-container/issues/5

Signed-off-by: Andreea Florescu <fandree@amazon.com>