update base image to ubuntu 22.04
While the latest available version is Ubuntu 23.04 at this point, we
cannot use it for this container because the kcov installation fails.

The problem is that the binutils changed the signature of
init_disassemble_info(). Until this is fixed in the kcov build there
is nothing we can do about it.

Signed-off-by: Andreea Florescu <fandree@amazon.com>