set default rust version in dockerfile directly
We were setting the rust version in the docker.sh instead.
We want to move to github workflows, and thus it makes more sense to
have the default in the Dockerfile as that's what is going to be used
by both docker.sh (which we will deprecated most likely) and the github
workflow.

Signed-off-by: Andreea Florescu <fandree@amazon.com>