reduce the size occupied during build by container
It was too large to be built as part of the free github runners.

To make it smaller a few things happened:
- use only one RUN command so that we don't have multiple container
layers. This ended up saving most of the space and allowing the build
to pass
- remove the Rust registry after installing binaries
- install the minimal Rust nightly toolchain as we need it just for
running the fuzzer
- remove the bindgen related tools that are not (yet) used
- remove pip3 after installing the packages we need
- do not use a cache when installing pip packages
- remove libgpio source files once installation is completed. For this
one I also did a bit of refactoring so that we use pushd & popd instead
of keeping track of directories.

Signed-off-by: Andreea Florescu <fandree@amazon.com>