Replace cargo-kcov with llvm-cov
cargo-kcov is no longer working as of rust toolchain version 1.71.0, as
it can no longer discover test executables. This is due to cargo
changing its verbose stdout printing to include the qualified path to
rustc, which causes cargo-kcov's string matching to fail.

See also https://github.com/kennytm/cargo-kcov/issues/54

cargo-kcov has had its last commit in 2019, so replace it with
cargo-llvm-cov, which is actively maintained, and written for rust.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>