fix: replaced ARCH with `uname -m`
In the previous commits the ARCH
env variable was introduced to the Dockerfile
and build_container.sh scripts. It was meant to
help with configuration of linux headers for
musl-tools, but did not work for some reason.
Considering the build_container.sh was already
using `uname -m` calls to get the current arch
this commit also uses it, so there is no need
for ARCH anymore.

Signed-off-by: Egor Lazarchuk <yegorlz@amazon.co.uk>