github: fix "Generate next docker tag" log output
The current output of the "Generate next docker tag" step is:

    Run echo "VERSION=$(./docker.sh print-next-version)" >> $GITHUB_ENV
    Next version to be published is:

The next version is not printed, because `$GITHUB_ENV` is loaded at
the next step. So at this point we cannot access variables that we
are setting in this step.

Use a temporary NEXT_VERSION variable where we save the next version
to print it correctly.

Signed-off-by: Stefano Garzarella <sgarzare@redhat.com>