Introduce riscv64 CI container
Add build scripts for v6.10 riscv64 kernel, qemu-system-riscv64 and
opensbi required to boot qemu-system inside docker container.

With this approach, we are able to run tests inside qemu-system, while
preserving the original output as much as possbile with ssh.

This is the third draft proceeds #101, #104. It is expected to be
replaced by the second draft #104 in the future which standardise
`riscv64`.

Signed-off-by: Ruoqing He <heruoqing@iscas.ac.cn>