Use --load with "docker build"
If --load is not provided to "docker build" command, the image only
remains in the local cache and is not available for use locally when
doing "docker run rustvmm/dev:$VERSION" like the README suggests.

Without this change using the image locally is not possible, the image
does not show up when doing `docker image ls` and the tag is not found
when doing `docker run <TAG>`. the `docker build` step prints this
warning: "WARNING: No output specified with docker-container driver.
Build result will only remain in the build cache."

Signed-off-by: Manos Pitsidianakis <manos.pitsidianakis@linaro.org>