docker.sh: Replace counter tagging with git sha1
We were using `vXX` counter tagging till `v49` for x86_64 and aarch64,
and `v49-riscv` for riscv64. And we have separated riscv64 from the
other two to another job due to insufficient Github Action time limit
(if the job failed to complete in 6 hours, it will be canceled), which
end up of having two jobs. If the x86_64 and aarch64 job are finished
and images are published in vN, while riscv64 job fails in that run,
restarting the riscv64 job will result an absence of vN and a presence
of v(N+1). So we decided to move on to git sha1 tagging strategy.

Remove `latest` function and `vXX` counter tagging, and use output of

```console
git show -s --format=%h
```

as the tag of images built per PR for publishing.

Signed-off-by: Ruoqing He <heruoqing@iscas.ac.cn>