ci: drop path filtering from pull_request target
Since the "docker container builds" sanity check is required for merging
(so that we can use auto-merge), it needs to always trigger, so drop the
filtering based on which files were modified.

There's the slight downside that it means we'll need to wait 3+ hours
even on PRs that dont actually touch the container, but those are very
rare from experience. If it ever becomes a nuisance, we can modify the
action to instead do the paths check inside the actual command step, and
return early.

Leave the check in the `push` target, so that we do not unnecessarily
build docker containers after merging though.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>