ci: use paths-ignore to determine when to trigger build
It's better to build too many docker containers than too few, so let's
go with not triggering the build if we know that only files irrelevant
to the container are modified (instead of trying to explicitly list all
files that are relevant). This way, if the list is stale, we build "too
many" containers, whereas with the current solution, we build "too few"
containers (see for example #132).

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>